+++ src/enclave/interface.h
@@ -46,8 +46,6 @@ struct CCFConfig
   consensus::Configuration consensus_config = {};
   ccf::NodeInfoNetwork node_info_network = {};
   size_t snapshot_tx_interval;
-  size_t max_open_sessions_soft;
-  size_t max_open_sessions_hard;
 
   // Only if joining or recovering
   std::vector<uint8_t> startup_snapshot;
@@ -103,8 +101,6 @@ DECLARE_JSON_REQUIRED_FIELDS(
   consensus_config,
   node_info_network,
   snapshot_tx_interval,
-  max_open_sessions_soft,
-  max_open_sessions_hard,
   startup_snapshot,
   startup_snapshot_evidence_seqno,
   signature_intervals,