+++ src/http/http_endpoint.h
@@ -17,7 +17,7 @@ namespace http
 
     HTTPEndpoint(
       http::Parser& p_,
-      size_t session_id,
+      int64_t session_id,
       ringbuffer::AbstractWriterFactory& writer_factory,
       std::unique_ptr<tls::Context> ctx) :
       TLSEndpoint(session_id, writer_factory, std::move(ctx)),
@@ -98,13 +98,13 @@ namespace http
     std::shared_ptr<enclave::RPCMap> rpc_map;
     std::shared_ptr<enclave::RpcHandler> handler;
     std::shared_ptr<enclave::SessionContext> session_ctx;
-    size_t session_id;
+    int64_t session_id;
     size_t request_index = 0;
 
   public:
     HTTPServerEndpoint(
       std::shared_ptr<enclave::RPCMap> rpc_map,
-      size_t session_id,
+      int64_t session_id,
       ringbuffer::AbstractWriterFactory& writer_factory,
       std::unique_ptr<tls::Context> ctx) :
       HTTPEndpoint(request_parser, session_id, writer_factory, std::move(ctx)),
@@ -223,7 +223,7 @@ namespace http
 
   public:
     HTTPClientEndpoint(
-      size_t session_id,
+      int64_t session_id,
       ringbuffer::AbstractWriterFactory& writer_factory,
       std::unique_ptr<tls::Context> ctx) :
       HTTPEndpoint(response_parser, session_id, writer_factory, std::move(ctx)),